catalog-backend: break out github org config for consistency
