+++ .changeset/good-news-return.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend': patch
+---
+
+Break out GithubOrgReaderProcessor config into its own file for consistency with the other org processors.
